pyproject.toml
[project]
name = "citation"
version = "0.1.0"
description = "Extract human-readable citation information from pdf, url, video and audio files using LLM"
authors = [
    { name = "ajia", email = "yyjfwoaini@gmail.com" }
]
dependencies = [
    "marker-pdf>=1.6.1",
    "dspy>=2.6.12",
    "pymupdf>=1.25.4",
    "pypdf>=3.0.0",
    "bibtexparser>=1.4.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.9.3",
    "requests>=2.25.1"
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
citation = "citation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=2.12.1"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["citation"]
